third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/ownerDocument-expected.txt

Test HTMLTemplateElement content ownerDocument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS template.ownerDocument is document
PASS template.content is content
PASS templateContentOwnerDocument instanceof HTMLDocument is true
PASS document is not templateContentOwnerDocument
PASS template.content.firstChild.ownerDocument is templateContentOwnerDocument
PASS innerTemplate.ownerDocument is templateContentOwnerDocument
PASS innerTemplate.content.ownerDocument is templateContentOwnerDocument
PASS innerTemplate.ownerDocument instanceof HTMLDocument is true
PASS template2.ownerDocument is document
PASS template2.content.ownerDocument is templateContentOwnerDocument
PASS template3.ownerDocument is document
PASS template3.content.ownerDocument is templateContentOwnerDocument
PASS template3.content.firstChild.ownerDocument is templateContentOwnerDocument
PASS successfullyParsed is true

TEST COMPLETE
